virt-handler, net: enter the virt-launcher's net-ns later
This commit first checks that there is actually an interface to plug and
just if there is one, enters the virt-launcher's network namespace.
